requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/base.in
#
amqp==2.6.1               # via kombu
argon2-cffi==20.1.0       # via -r requirements/base.in
asgiref==3.2.10           # via django
billiard==3.6.3.0         # via celery
boto3==1.14.40            # via -r requirements/base.in
botocore==1.17.40         # via boto3, s3transfer
cachetools==4.1.1         # via google-auth
celery[redis]==4.4.7      # via -r requirements/base.in
certifi==2020.6.20        # via requests
cffi==1.14.1              # via argon2-cffi, cryptography
chardet==3.0.4            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
cryptography==3.0         # via -r requirements/base.in
defusedxml==0.6.0         # via python3-openid
django-allauth==0.42.0    # via django-rest-auth
django-cors-headers==3.4.0  # via -r requirements/base.in
django-environ==0.4.5     # via -r requirements/base.in
django-filter==2.3.0      # via -r requirements/base.in
django-model-utils==4.0.0  # via -r requirements/base.in
django-redis==4.12.1      # via -r requirements/base.in
django-rest-auth[with_social]==0.9.5  # via -r requirements/base.in
django-storages==1.9.1    # via -r requirements/base.in
django==3.0.6             # via -r requirements/base.in, django-allauth, django-cors-headers, django-filter, django-model-utils, django-redis, django-rest-auth, django-storages, djangorestframework, drf-yasg
djangorestframework==3.11.1  # via -r requirements/base.in, django-rest-auth, drf-yasg
docutils==0.15.2          # via botocore
drf-yasg==1.17.1          # via -r requirements/base.in
google-api-core==1.22.0   # via google-api-python-client
google-api-python-client==1.10.0  # via -r requirements/base.in
google-auth-httplib2==0.0.4  # via google-api-python-client
google-auth==1.20.1       # via google-api-core, google-api-python-client, google-auth-httplib2
googleapis-common-protos==1.52.0  # via google-api-core
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                # via requests
inflection==0.5.0         # via drf-yasg
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
jmespath==0.10.0          # via boto3, botocore
kombu==4.6.11             # via celery
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.4           # via -r requirements/base.in, drf-yasg
protobuf==3.12.4          # via google-api-core, googleapis-common-protos
psycopg2==2.8.3
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via -r requirements/base.in
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via -r requirements/base.in, botocore
python3-openid==3.2.0     # via django-allauth
pytz==2020.1              # via celery, django, google-api-core
redis==3.5.3              # via celery, django-redis
requests-oauthlib==1.3.0  # via django-allauth
requests==2.24.0          # via coreapi, django-allauth, google-api-core, requests-oauthlib
rsa==4.6                  # via google-auth
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via drf-yasg
s3transfer==0.3.3         # via boto3
six==1.15.0               # via argon2-cffi, cryptography, django-rest-auth, drf-yasg, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, packaging, protobuf, python-dateutil
sqlparse==0.3.1           # via django
uritemplate==3.0.1        # via coreapi, drf-yasg, google-api-python-client
urllib3==1.25.10          # via botocore, requests
vine==1.3.0               # via amqp, celery
whitenoise==5.2.0         # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
